Update scroll transition function
