+++ src/home.js
@@ -4,7 +4,7 @@ import about from './about.js';
 
 export default function homePage() {
   const mainContainer = document.querySelector('#content');
-  mainContainer.innerText = '';
+  mainContainer.innerHTML = '';
   mainContainer.appendChild(navbar());
   mainContainer.appendChild(header());
   mainContainer.appendChild(about());
